cells#search now returns era in json object
this is somewhat of a workaround, since era is passed into cells#search initially. I'd argue, though, that this is less coupled than the alternative. The updateResults function needs certain information to work. Instead of getting some of that information from JS and some from the JSON response, now all information comes from the JSON response. If it's going to couple to something, lets make it couple to only one thing.